+++ CHANGES.rst
@@ -9,6 +9,7 @@ Changelog
   [folix-01]
  
 - Bypass the today delete limit for the 'out-of-office' types
+- Add manager notification on booking canceled
   [folix-01]
 
 - Change the booking notification flag label.